cummit

+++ Animation.h
@@ -2,11 +2,11 @@
 class Animation
 {
 public:
-	short unsigned int keyCount;
+	unsigned int keyCount;
 	float* timeKeys;
 	float* stateKeys;
 
-	Animation(short unsigned int keyCount, float* timeKeys, float* stateKeys)
+	Animation(unsigned int keyCount, float* timeKeys, float* stateKeys)
 	{
 		this->keyCount = keyCount;
 		this->timeKeys = timeKeys;
